feat(tslint): change plugin API
This commit changes the API to load the TSLint results JSON file from disk instead of using TSLint's
programmatic API to run TSLint. While this requires an extra setup step for users, this allows this
plugin to be functional (v1.0.0 did not work properly).

BREAKING CHANGE: there is a new configuration API, so this will result in a major version bump.